// src/app/interfaces/Empresa.ts
export interface Empresa {
  id:             number;
  ciudad:         string;
  codigoPostal:   number;
  colonia:        string;
  correo:         string;
  direccionCalle: string;
  logo:           File;
  logoPath:       string;
  moneda:         string;
  nombre:         string;
  numeroExterior: string;
  numeroPuerta:   string;
  pais:           string;
  provincia:      string;
  registro:       string;
  rfc:            string;
  sitioWeb:       string;
  telefono:       string;
}